# Optimize postings offset table reading