@colega 2022-11-11
Thank you for your feedback, @bboreham, I applied it, also changing the way label & value are passed to the func: they're []byte now. 

Updated the benchmark in the description PTAL.